Fix review findings on badges: error-state copy, singular units, close button
- BadgesDialog: only show the "n av n låst opp" subtitle when there are
  badges to count, so a failed fetch no longer shows a false "0 av 0"
  right above the error message; add a Lukk button via DialogActions to
  match the other dialogs in the app.
- badges.js: use singular Norwegian units ("1 klubb"/"1 fylke") instead
  of plural when a tier's target is 1.
- badges.test.js / BadgesDialog.test.js: add the missing assertions
  (error path has no "låst opp" text, eleven badges in group order,
  computeBadges({}) is safe) and update the two copy assertions the
  fixes above change.
- DashboardStats.roundCount: int -> long, matching its neighbours.
- Update CLAUDE.md and the badges design doc to match what shipped.

+++ backend/src/main/java/fritids/norskgolf/dto/DashboardStats.java
@@ -15,7 +15,7 @@ public class DashboardStats {
     private String displayName;
     private String avatar;
     private String email;
-    private int roundCount;
+    private long roundCount;
 
     // --- Inner Class: Round Summary ---
     public static class RoundSummary {
@@ -59,8 +59,8 @@ public class DashboardStats {
 
     // Every round ever logged. recentRounds is capped at 5, so this is the only
     // lifetime total the frontend can see.
-    public int getRoundCount() { return roundCount; }
-    public void setRoundCount(int roundCount) { this.roundCount = roundCount; }
+    public long getRoundCount() { return roundCount; }
+    public void setRoundCount(long roundCount) { this.roundCount = roundCount; }
 
     public long getTotalPlayed() { return totalPlayed; }
     public void setTotalPlayed(long totalPlayed) { this.totalPlayed = totalPlayed; }
